+++ packages/ub-pub/ClientRepository.js
@@ -195,6 +195,7 @@ class ClientRepository extends CustomRepository {
 
   /**
    * Alias to {@link class:ClientRepository#selectAsObject selectAsObject}
+   * @return {Promise<Array<object>>}
    */
   select (fieldAliases) {
     return this.selectAsObject(fieldAliases)
@@ -210,7 +211,7 @@ class ClientRepository extends CustomRepository {
    * @return {Promise<object|undefined>} Promise, resolved to {Object|undefined}
    */
   selectSingle (fieldAliases) {
-    return this.selectAsObject(fieldAliases).then(function (rows) {
+    return this.selectAsObject(fieldAliases).then((rows) => {
       if (rows.length > 1) console.error(this.CONSTANTS.selectSingleMoreThanOneRow)
       return rows[0]
     })
@@ -223,7 +224,7 @@ class ClientRepository extends CustomRepository {
    * @return {Promise<Number|String|undefined>} Promise, resolved to {Number|String|undefined}
    */
   selectScalar () {
-    return this.selectAsArray().then(function (result) {
+    return this.selectAsArray().then((result) => {
       const L = result.resultData.data.length
       if (L) {
         if (L > 1) console.error(this.CONSTANTS.selectScalarMoreThanOneRow)